requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM APIs
openai>=1.0.0
anthropic>=0.23.0
google-generativeai>=0.5.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Optional for async operations
aiohttp>=3.8.0